refactor(gui): inline open_in_file_manager into utils
Drop the dedicated src/utils/open_path.rs module — a single ~15 line
helper does not warrant its own file. Move the function (still using
cfg_select!) into src/utils.rs and update both call sites accordingly.

Refs #105
